Snore_Detection/requirements.txt
numpy
pandas
scikit-learn
librosa
tensorflow
keras
matplotlib
pytorch
pytorch_lightning
pyaudio
pyserial
